raspberry/requirements.txt
Pygments
RPi.GPIO
